+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+WIP
+
+## [0.1.4] - 2025-07-23
+
 ### Fixed
 
 - Removed .git/sup.lock file when execution is terminated from Ctrl+C.
@@ -79,7 +83,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Basic functionality to stash uncommitted changes, pull from remote, and restore changes.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/strowk/sup/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/strowk/sup/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/strowk/sup/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/strowk/sup/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/strowk/sup/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/strowk/sup/compare/v0.1.0...v0.1.1